requirements.txt
asgiref==3.11.0
certifi==2025.11.12
charset-normalizer==3.4.4
distlib==0.4.0
Django==6.0
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
filelock==3.20.2
gunicorn==23.0.0
idna==3.11
packaging==25.0
platformdirs==4.5.1
psycopg2==2.9.11
PyJWT==2.10.1
requests==2.32.5
sqlparse==0.5.5
tzdata==2025.3
urllib3==2.6.2
virtualenv==20.35.4
